+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -467,13 +467,3 @@ private extension UIButton {
         }
     }
 }
-
-private extension UIRefreshControl {
-    func simulatePullToRefresh() {
-        allTargets.forEach { target in
-            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach {
-                (target as NSObject).perform(Selector($0))
-            }
-        }
-    }
-}
